+++ common/src/lib.rs
@@ -6,7 +6,7 @@ pub mod utils;
 pub use address::{Address, AddressPayload, AddressType, CodeHashIndex};
 pub use {anyhow, anyhow::Result, async_trait::async_trait, derive_more, minstant};
 
-use ckb_types::{h256, H256};
+use ckb_types::{bytes::Bytes, core::BlockNumber, h256, packed, H256};
 use derive_more::Display;
 use serde::{Deserialize, Serialize};
 
@@ -250,6 +250,23 @@ impl<T> Default for PaginationResponse<T> {
     fn default() -> Self {
         Self::new(vec![])
     }
+}
+
+#[derive(Clone, Debug, Hash, PartialEq, Eq)]
+pub struct DetailedCell {
+    pub epoch_number: u64,
+    pub block_number: BlockNumber,
+    pub block_hash: H256,
+    pub tx_index: u32,
+    pub out_point: packed::OutPoint,
+    pub cell_output: packed::CellOutput,
+    pub cell_data: Bytes,
+    pub consumed_block_number: Option<u64>,
+    pub consumed_block_hash: Option<H256>,
+    pub consumed_tx_hash: Option<H256>,
+    pub consumed_tx_index: Option<u32>,
+    pub consumed_input_index: Option<u32>,
+    pub since: Option<u64>,
 }
 
 pub fn display_list_as_hex<T: AsRef<[u8]>>(list: &[T]) {
