+++ core/rpc/core/src/impl/utils_types.rs
@@ -1,6 +1,5 @@
-use common::PaginationRequest;
+use common::{DetailedCell, PaginationRequest};
 use core_rpc_types::{AssetInfo, Item};
-use core_storage::DetailedCell;
 
 use ckb_types::packed;
